Update in the layout method.
