+++ Example/CrownControl/PDF/PDFViewController.swift
@@ -43,10 +43,10 @@ class PDFViewController: UIViewController {
         crownViewController = CrownIndicatorViewController(with: attributes)
         
         // Cling the bottom of the crown to the bottom of the web view with -35 offset
-        let verticalConstraint = CrownAxisConstraint(crownEdge: .bottom, anchorView: webView, anchorViewEdge: .bottom, offset: -35)
+        let verticalConstraint = CrownAttributes.AxisConstraint(crownEdge: .bottom, anchorView: webView, anchorViewEdge: .bottom, offset: -35)
         
         // Cling the bottom of the crown to the bottom of its superview with -50 offset
-        let horizontalConstraint = CrownAxisConstraint(crownEdge: .trailing, anchorView: view, anchorViewEdge: .trailing, offset: -50)
+        let horizontalConstraint = CrownAttributes.AxisConstraint(crownEdge: .trailing, anchorView: view, anchorViewEdge: .trailing, offset: -50)
         
         crownViewController.layout(in: self, horizontalConstaint: horizontalConstraint, verticalConstraint: verticalConstraint)
     }
