+++ Example/CrownControl/Photos/PhotoCollectionViewController.swift
@@ -57,10 +57,10 @@ class PhotoCollectionViewController: UIViewController {
         crownViewController = CrownIndicatorViewController(with: attributes, delegate: self)
         
         // Cling the bottom of the crown to the bottom of the web view with -35 offset
-        let verticalConstraint = CrownAxisConstraint(crownEdge: .bottom, anchorView: view, anchorViewEdge: .bottom, offset: -70)
+        let verticalConstraint = CrownAttributes.AxisConstraint(crownEdge: .bottom, anchorView: view, anchorViewEdge: .bottom, offset: -70)
         
         // Cling the bottom of the crown to the bottom of its superview with -50 offset
-        let horizontalConstraint = CrownAxisConstraint(crownEdge: .centerX, anchorView: view, anchorViewEdge: .centerX)
+        let horizontalConstraint = CrownAttributes.AxisConstraint(crownEdge: .centerX, anchorView: view, anchorViewEdge: .centerX)
         
         crownViewController.layout(in: self, horizontalConstaint: horizontalConstraint, verticalConstraint: verticalConstraint)
     }
