+++ Source/CrownAttributes/CrownAttributes+AxisConstraint.swift
@@ -1,0 +1,25 @@
+//
+//  CrownAttributes+AxisConstraint.swift
+//  CrownControl
+//
+//  Created by Daniel Huri on 11/24/18.
+//
+
+import UIKit
+import QuickLayout
+
+public extension CrownAttributes {
+    public struct AxisConstraint {
+        public let crownEdge: QLAttribute
+        public let anchorView: UIView
+        public let anchorViewEdge: QLAttribute
+        public let offset: CGFloat
+        
+        public init(crownEdge: QLAttribute, anchorView: UIView, anchorViewEdge: QLAttribute, offset: CGFloat = 0) {
+            self.crownEdge = crownEdge
+            self.anchorView = anchorView
+            self.anchorViewEdge = anchorViewEdge
+            self.offset = offset
+        }
+    }
+}
